requirements.txt
langchain
datasets
pandas
numpy
python-dotenv
openai
pymupdf
streamlit